refactor: update dto 수정
