+++ backend/chat-server/src/main/java/com/smiletogether/chatserver/dto/request/ChannelMessageUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.smiletogether.chatserver.service.dto;
+package com.smiletogether.chatserver.dto.request;
 
 import java.time.LocalDateTime;
 
@@ -6,7 +6,7 @@ public record ChannelMessageUpdateRequest(
         String type,
         String messageId,
         String content,
-        LocalDateTime updateAt
+        LocalDateTime createdAt
 ) {
 
 }
